Added clock tests for metrics
