+++ android-test/app/src/test/java/co/elastic/apm/android/test/attributes/traces/GlobalAttributeTest.java
@@ -1,18 +1,11 @@
 package co.elastic.apm.android.test.attributes.traces;
 
-import static org.junit.Assert.assertTrue;
-import static org.mockito.Mockito.verify;
-
 import org.junit.Test;
 
 import java.util.List;
 
-import co.elastic.apm.android.sdk.internal.time.ntp.NtpManager;
-import co.elastic.apm.android.sdk.traces.common.tools.ElasticTracer;
 import co.elastic.apm.android.test.common.spans.Spans;
-import co.elastic.apm.android.test.testutils.TestElasticClock;
 import co.elastic.apm.android.test.testutils.base.BaseRobolectricTest;
-import io.opentelemetry.api.trace.Span;
 import io.opentelemetry.sdk.trace.data.SpanData;
 
 public class GlobalAttributeTest extends BaseRobolectricTest {
@@ -31,43 +24,6 @@ public class GlobalAttributeTest extends BaseRobolectricTest {
 
         Spans.verify(customSpan)
                 .hasAttribute("type", "mobile");
-    }
-
-    @Test
-    public void whenASpanIsCreated_itHasTimestampSetFromElasticClock() {
-        long startTimeFromElasticClock = 123456789;
-        NtpManager ntpManager = getAgentDependenciesProvider().getNtpManager();
-        TestElasticClock clock = (TestElasticClock) ntpManager.getClock();
-        clock.setForcedNow(startTimeFromElasticClock);
-        SpanData span = getSpanData();
-
-        Spans.verify(span)
-                .startedAt(startTimeFromElasticClock);
-    }
-
-    @Test
-    public void whenClockNowChangesInMidSpan_verifyFinalSpanDurationIsNotAffected() {
-        long startTimeFromElasticClock = 2_000_000_000;
-        NtpManager ntpManager = getAgentDependenciesProvider().getNtpManager();
-        TestElasticClock clock = (TestElasticClock) ntpManager.getClock();
-        clock.setForcedNow(startTimeFromElasticClock);
-
-        Span span = ElasticTracer.androidActivity().spanBuilder("TimeChangeSpan").startSpan();
-
-        // Moving now backwards:
-        clock.setForcedNow(1_000_000_000L);
-
-        span.end();
-
-        SpanData recordedSpan = getRecordedSpan();
-        Spans.verify(recordedSpan)
-                .startedAt(startTimeFromElasticClock);
-        assertTrue(recordedSpan.getEndEpochNanos() > startTimeFromElasticClock);
-    }
-
-    @Test
-    public void whenASpanIsCreated_itHasInitializedTheNtpManager() {
-        verify(getAgentDependenciesProvider().getNtpManager()).initialize();
     }
 
     private SpanData getSpanData() {
